Add files via upload
I currently have the accounting data in a google sheets. The cashtally_v7 program gets the user input of the cash tally, and the denominations, and calculates the total. Then it finally saves the record in the google sheets. I merged the google sheets with a firebase real time database, successfully, but now I want to retrieve that data to display it on my html webpage and perform search operations on it by taking user input of data. For some reason, I am not able to view the firebase data on my html page. Can anyone assist me on this?